+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,39 @@
+---
+name: Bug Report
+about: Create a report to help us improve
+title: ""
+labels: "bug"
+assignees: ""
+---
+
+**Please review our [Contributing Guide](CONTRIBUTING.md) before opening an issue.**
+
+### Description
+
+A clear and concise description of what the bug is.
+
+### To Reproduce
+
+Please provide the specific command(s) you ran:
+
+```sh
+eka ...
+```
+
+### Expected Behavior
+
+A clear and concise description of what you expected to happen.
+
+### Actual Behavior
+
+Please paste the full output of the command, including any error messages.
+
+### Environment
+
+- **OS:** [e.g., Linux, macOS, Windows]
+- **`eka` Version:** [Run `eka --version`]
+- **Shell:** [e.g., bash, zsh, fish]
+
+### Additional Context
+
+Add any other context about the problem here.
