+++ adrs/0000-template.md
@@ -1,0 +1,19 @@
+# [NUMBER]. [TITLE]
+
+Date: [YYYY-MM-DD]
+
+## Status
+
+[Proposed | Accepted | Deprecated | Superseded]
+
+## Context
+
+A detailed explanation of the issue, the context in which it arises, and the problem that needs to be solved. This section should provide enough information for someone new to the project to understand the forces at play.
+
+## Decision
+
+A clear and concise description of the solution that has been decided upon. This should explain how the solution addresses the problem outlined in the context.
+
+## Consequences
+
+A description of the results of the decision. This should include both the positive outcomes and any potential drawbacks or trade-offs. It should also outline any new issues that might arise from this decision.
