+++ crates/atom/src/manifest.rs
@@ -6,14 +6,14 @@
 //! ## Manifest Structure
 //!
 //! Every atom must have a manifest file named `atom.toml` that contains at minimum
-//! an `[atom]` section with the atom's ID, version, and optional description.
+//! a `[atom]` section with the atom's ID, version, and optional description.
 //! Additional sections can specify dependencies and other configuration.
 //!
 //! ## Key Types
 //!
-//! - [`Manifest`] - The complete manifest structure
-//! - [`Atom`] - The core atom metadata (id, version, description)
-//! - [`AtomError`] - Errors that can occur during manifest processing
+//! - [`Manifest`] - The complete manifest structure, representing the `atom.toml` file.
+//! - [`Atom`] - The core atom metadata (`tag`, `version`, `description`).
+//! - [`AtomError`] - Errors that can occur during manifest processing.
 //!
 //! ## Example Manifest
 //!
@@ -38,28 +38,31 @@
 //!
 //! ## Usage
 //!
+//! Manifests can be created programmatically or parsed from a string or file.
+//!
 //! ```rust,no_run
+//! use std::str::FromStr;
+//!
 //! use atom::manifest::Manifest;
 //! use atom::{Atom, AtomTag};
 //! use semver::Version;
 //!
-//! // Create a manifest programmatically
+//! // Create a manifest programmatically.
 //! let manifest = Manifest::new(
 //!     AtomTag::try_from("my-atom").unwrap(),
 //!     Version::new(1, 0, 0),
 //!     Some("My first atom".to_string()),
 //! );
 //!
-//! // Parse a manifest from a string
+//! // Parse a manifest from a string.
 //! let manifest_str = r#"
 //! [atom]
 //! tag = "parsed-atom"
 //! version = "2.0.0"
 //! "#;
-//! let parsed: Manifest = manifest_str.parse().unwrap();
+//! let parsed = Manifest::from_str(manifest_str).unwrap();
 //! ```
 
-pub mod deps;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::str::FromStr;
@@ -69,41 +72,44 @@ use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use toml_edit::{DocumentMut, de};
 
+use crate::id::Name;
 use crate::{Atom, AtomTag};
 
-/// Errors which occur during manifest (de)serialization.
+pub mod deps;
+
+/// A specialized result type for manifest operations.
+pub type AtomResult<T> = Result<T, AtomError>;
+
+/// An error that can occur when parsing or handling an atom manifest.
 #[derive(Error, Debug)]
 pub enum AtomError {
-    /// The manifest is missing the required \[atom] key.
+    /// The manifest is missing the required `[atom]` table.
     #[error("Manifest is missing the `[atom]` key")]
     Missing,
-    /// One of the fields in the required \[atom] key is missing or invalid.
+    /// One of the fields in the `[atom]` table is missing or invalid.
     #[error(transparent)]
     InvalidAtom(#[from] de::Error),
     /// The manifest is not valid TOML.
     #[error(transparent)]
     InvalidToml(#[from] toml_edit::TomlError),
-    /// The manifest could not be read.
+    /// An I/O error occurred while reading the manifest file.
     #[error(transparent)]
     Io(#[from] std::io::Error),
 }
 
-type AtomResult<T> = Result<T, AtomError>;
-use crate::id::Name;
-
-/// The type representing the required fields of an Atom's manifest.
+/// Represents the structure of an `atom.toml` manifest file.
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
 #[serde(deny_unknown_fields)]
 pub struct Manifest {
-    /// The required \[atom] key of the TOML manifest.
+    /// The required `[atom]` table, containing core metadata.
     pub atom: Atom,
-    /// The dependencies of the Atom.
+    /// The dependencies of the atom.
     #[serde(default, skip_serializing_if = "HashMap::is_empty")]
     pub(crate) deps: HashMap<Name, deps::Dependency>,
 }
 
 impl Manifest {
-    /// Create a new atom Manifest with the given values.
+    /// Creates a new `Manifest` with the given tag, version, and description.
     pub fn new(tag: AtomTag, version: Version, description: Option<String>) -> Self {
         Manifest {
             atom: Atom {
@@ -115,13 +121,13 @@ impl Manifest {
         }
     }
 
-    /// Build an Atom struct from the \[atom] key of a TOML manifest,
-    /// ignoring other fields or keys].
+    /// Parses an [`Atom`] struct from the `[atom]` table of a TOML document string,
+    /// ignoring other tables and fields.
     ///
     /// # Errors
     ///
-    /// This function will return an error if the content is invalid
-    /// TOML, or if the \[atom] key is missing.
+    /// This function will return an error if the content is invalid TOML,
+    /// or if the `[atom]` table is missing.
     pub(crate) fn get_atom(content: &str) -> AtomResult<Atom> {
         let doc = content.parse::<DocumentMut>()?;
