+++ src/cli/commands/resolve/mod.rs
@@ -1,25 +1,31 @@
+//! This module defines the `resolve` subcommand.
+//!
+//! The `resolve` subcommand is responsible for resolving dependencies for a
+//! given set of atoms and writing the results to a lock file.
+
 use std::path::PathBuf;
 
 use anyhow::Result;
 use clap::Parser;
 
-mod git;
-
 use crate::cli::store::Detected;
 
+mod git;
+
+/// The `resolve` subcommand.
 #[derive(Parser, Debug)]
 #[command(arg_required_else_help = true)]
 pub struct Args {
-    /// The path of the atom(s) to resolve dependencies for
+    /// The path of the atom(s) to resolve dependencies for.
     path: Vec<PathBuf>,
-    #[command(flatten)]
-    store: StoreArgs,
-    /// Output file for the lock (default: atom.lock)
+    /// The output file for the lock (default: `atom.lock`).
     #[arg(short, long, default_value = "atom.lock")]
     output: PathBuf,
-    /// Resolution mode: shallow or deep (default: shallow)
+    /// The resolution mode: `shallow` or `deep` (default: `shallow`).
     #[arg(short, long, default_value = "shallow")]
     mode: String,
+    #[command(flatten)]
+    store: StoreArgs,
 }
 
 #[derive(Parser, Debug)]
@@ -28,6 +34,7 @@ struct StoreArgs {
     git: git::GitArgs,
 }
 
+/// The main entry point for the `resolve` subcommand.
 pub(super) fn run(_store: Detected, _args: Args) -> Result<()> {
     Ok(())
 }
