+++ src/cli/store.rs
@@ -1,18 +1,37 @@
+//! This module handles the detection of the underlying version control system.
+
+use std::fs;
+
 use atom::store::git;
 use gix::ThreadSafeRepository;
 use thiserror::Error;
 
+/// Represents the detected version control system.
 #[non_exhaustive]
 #[derive(Clone, Debug)]
 pub(super) enum Detected {
+    /// A Git repository was detected.
     Git(&'static ThreadSafeRepository),
+    /// No supported version control system was detected.
     #[allow(dead_code)]
     None,
 }
 
+/// Errors that can occur during repository detection.
+#[derive(Error, Debug)]
+pub(crate) enum Error {
+    /// No supported repository was found in the current directory or its parents.
+    #[error("No supported repository found in this directory or its parents")]
+    FailedDetection,
+
+    /// An error occurred while discovering the repository.
+    #[error(transparent)]
+    Discover(#[from] Box<gix::discover::Error>),
+}
+
+/// Detects the version control system in the current directory.
 pub(super) async fn detect() -> Result<Detected, Error> {
     if let Ok(Some(repo)) = git::repo() {
-        use std::fs;
         let git_dir = fs::canonicalize(repo.path())
             .ok()
             .map(|p| p.display().to_string());
@@ -26,13 +45,4 @@ pub(super) async fn detect() -> Result<Detected, Error> {
     }
 
     Err(Error::FailedDetection)
-}
-
-#[derive(Error, Debug)]
-pub(crate) enum Error {
-    #[error("No supported repository found in this directory or its parents")]
-    FailedDetection,
-
-    #[error(transparent)]
-    Discover(#[from] Box<gix::discover::Error>),
 }
