+++ src/lib.rs
@@ -1,1 +1,5 @@
+//! Eka, a simple, fast, and experimental package manager for Nix.
+
+#![warn(missing_docs)]
+
 pub mod cli;
