+++ src/main.rs
@@ -1,8 +1,13 @@
+//! The main entry point for the Eka CLI.
+
+#![warn(missing_docs)]
+
 use std::process::ExitCode;
 
 use clap::Parser;
 use eka::cli::{self, Args};
 
+/// The main entry point for the Eka CLI.
 #[tokio::main]
 async fn main() -> ExitCode {
     let args = Args::parse_from(cli::change_directory());
